feat: add `--serve` option
Add option to open a websocket server and listen for requests to either replace the currently running phonebook with a transmitted one or to reset the current phonebook.
Added a minimal webpage `fernspielctl_test.html` that can be used for basic testing of the protocol.
This required larger changes to `main.rs`, where some functionality has been pulled out into the new `app` module
and several other smaller modules.

closes #7