chore: add `expensive_tests` feature
Ignore tests that first need to estimate lag unless the new feature `expensive_tests` is set on the test build. Add a commented line to `Cargo.toml` that makes this a default feature in order to always run the expensive tests until uncommented again. Some other tests still take more than a second, but the now ignored ones are worse.