doc: add and update links in README
Update the introduction link in the README to a link to the phonebook format specification. Also add links to the _fernspieleditor_ and _weichspielapparat_ projects. Explicitly point out that self-compiling is not required, and point to the binary releases page. Also add a note about environment variables on Mac. Make note of the `--serve` command line option and the default port.`